casting int to float

+++ gfx/display_servers/dispserv_win32.c
@@ -107,7 +107,7 @@ static void win32_display_server_destroy(void *data)
    
    if (win32_orig_width > 0 && win32_orig_height > 0 )
       video_display_server_switch_resolution(win32_orig_width, win32_orig_height,
-         win32_orig_refresh , win32_orig_refresh );
+         win32_orig_refresh , (float)win32_orig_refresh );
 
 #ifdef HAS_TASKBAR_EXT
    if (g_taskbarList && win32_taskbar_is_created())
